+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 ## [Unreleased]
 
+## [0.3.0] - 2024-07-30
+
+- on_inner_monologue: ->(text:) { puts Rainbow("monologue: #{text}").yellow }
+- on_outer_voice: ->(text:) { puts Rainbow("voice: #{text}").green }
+- on_action_request: ->(text:) { puts Rainbow("action: #{text}").red }
+- on_pack_history: ->(text:) { puts Rainbow("summary: #{text}").blue }
+
 ## [0.2.5] - 2024-07-30
 
 - Improved start template